# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "Node.js API model",
    "description": "Node.js Express API Microservice with Mongo",
    "contact": {
      "email": "david.a.m@live.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080/api"
    },
    {
      "url": "https://localhost:8080/api"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "description": "Get all users",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No users found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      },
      "post": {
        "tags": ["Users"],
        "summary": "Add a new user to the store",
        "description": "Add a new user to the store",
        "operationId": "postUsers",
        "requestBody": {
          "description": "Create a new user in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get users by ID",
        "description": "Find a user searching by ID",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The MongoID of the user to be fetched",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete an existing user",
        "description": "Delete an existing user by Id",
        "operationId": "deleteUsers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The MongoID of the user to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user ID supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "User ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update an existing user",
        "description": "Update an existing user by Id",
        "operationId": "updateUsers",
        "requestBody": {
          "description": "Update an existent user in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The MongoID of the user to be fetched",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid MongoID supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "There is no user with that MongoID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "NewUser": {
        "required": ["name", "password", "role", "email"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Smith"
          },
          "password": {
            "type": "string",
            "example": "passw0rd!"
          },
          "email": {
            "type": "string",
            "example": "j.smith@gmail.com"
          },
          "role": {
            "type": "string",
            "example": "ADMIN_ROLE",
            "enum": ["ADMIN_ROLE", "USER_ROLE"]
          }
        }
      },
...
      "User": {
        "required": ["name", "category"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24,
            "example": "68373827924ffbeeabc4b1ed"
          },
          "name": {
            "type": "string",
            "example": "John Smith"
          },
          "email": {
            "type": "string",
            "example": "test3@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "12345678"
          },
          "img": {
            "type": "string",
            "example": "test.jpg"
          },
          "role": {
            "type": "string",
            "example": "ADMIN_ROLE",
            "enum": ["ADMIN_ROLE", "USER_ROLE"]
          },
          "google": {
            "type": "boolean",
            "description": "has Google authentication"
          },
          "status": {
            "type": "boolean",
            "description": "current status in the store"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "BadRequest"
          },
          "title": {
            "type": "string",
            "example": "Invalid ID"
          },
          "detail": {
            "type": "string",
            "example": "The ID supplied is not a valid MongoID value"
          }
        }
      }
    },
    "securitySchemes": {
      "token_auth": {
        "type": "apiKey",
        "name": "x-token",
        "in": "header"
      }
    }
  }
}